finalizado mvvm
