+++ app/src/main/res/values/strings.xml
@@ -16,4 +16,5 @@
     <string name="campo_obrigatorio">Campo obrigatório!</string>
     <string name="pratos_principais">Pratos Principais</string>
 
+
 </resources>
